requirements.txt
asgiref==3.4.1
blis==0.7.5
catalogue==2.0.6
certifi==2021.10.8
charset-normalizer==2.0.9
click==7.1.2
colorama==0.4.4
cymem==2.0.6
Django==4.0
django-cors-headers==3.10.1
djangorestframework==3.13.1
en-ner-bc5cdr-md==0.4.0
idna==3.3
Jinja2==3.0.3
joblib==1.1.0
langcodes==3.3.0
MarkupSafe==2.0.1
murmurhash==1.0.6
mysql==0.0.3
mysqlclient==2.1.0
numpy==1.21.5
packaging==21.3
pandas==1.3.5
pathy==0.6.1
preshed==3.0.6
psutil==5.8.0
pybind11==2.6.1
pydantic==1.8.2
PyMySQL==1.0.2
pyparsing==3.0.6
pysbd==0.3.4
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
requests==2.26.0
scikit-learn==1.0.2
scipy==1.7.3
six==1.16.0
smart-open==5.2.1
spacy==3.0.7
spacy-legacy==3.0.8
spacy-loggers==1.0.1
sqlparse==0.4.2
srsly==2.4.2
thinc==8.0.13
threadpoolctl==3.0.0
tqdm==4.62.3
typer==0.3.2
typing_extensions==4.0.1
tzdata==2021.5
urllib3==1.26.7
wasabi==0.9.0
